usePrefs hook"
